+++ data-processing-lib/python/src/data_processing/runtime/pure_python/__init__.py
@@ -1,7 +1,7 @@
 from data_processing.runtime.pure_python.runtime_configuration import PythonTransformRuntimeConfiguration
 from data_processing.runtime.pure_python.execution_configuration import PythonTransformExecutionConfiguration
-from data_processing.runtime.pure_python.transform_file_processor import PythonTransformFileProcessor
-from data_processing.runtime.pure_python.pooled_runtime_file_processor import process_transforms
+from data_processing.runtime.pure_python.transform_file_processor import (PythonTransformFileProcessor,
+                                                                          PythonPoolTransformFileProcessor)
 from data_processing.runtime.pure_python.transform_orchestrator import orchestrate
 from data_processing.runtime.pure_python.transform_launcher import PythonTransformLauncher
 from data_processing.runtime.pure_python.transform_invoker import invoke_transform, execute_python_transform
