+++ data-processing-lib/python/src/data_processing/runtime/pure_python/transform_orchestrator.py
@@ -11,15 +11,16 @@
 ################################################################################
 
 import time
+from multiprocessing import Pool
 import traceback
 from datetime import datetime
 
 from data_processing.data_access import DataAccessFactoryBase
 from data_processing.runtime.pure_python import (
-    PythonTransformFileProcessor,
     PythonTransformExecutionConfiguration,
     PythonTransformRuntimeConfiguration,
-    process_transforms,
+    PythonTransformFileProcessor,
+    PythonPoolTransformFileProcessor,
 )
 from data_processing.transform import TransformStatistics
 from data_processing.utils import get_logger
@@ -64,31 +65,12 @@ def orchestrate(
             print_interval = 1
         if execution_config.multiprocessing:
             # using multiprocessor pool for execution
-            statistics = process_transforms(files=files, size=execution_config.num_processors,
-                                            data_access_factory=data_access_factory, print_interval=print_interval,
-                                            runtime_configuration=runtime_config)
+            statistics = _process_transforms_multiprocessor(files=files, size=execution_config.num_processors,
+                                                           data_access_factory=data_access_factory, print_interval=print_interval,
+                                                           runtime_configuration=runtime_config)
         else:
-            # create executor
-            executor = PythonTransformFileProcessor(
-                data_access_factory=data_access_factory, statistics=statistics, runtime_configuration=runtime_config
-            )
-            # process data
-            logger.debug(f"{runtime_config.get_name()} Begin processing files")
-            t_start = time.time()
-            completed = 0
-            for path in files:
-                executor.process_file(path)
-                completed += 1
-                if completed % print_interval == 0:
-                    logger.info(
-                        f"Completed {completed} files ({100 * completed / len(files)}%) "
-                        f"in {(time.time() - t_start)/60} min"
-                    )
-            logger.info(f"Done processing {completed} files, waiting for flush() completion.")
-            # invoke flush to ensure that all results are returned
-            start = time.time()
-            executor.flush()
-            logger.info(f"done flushing in {time.time() - start} sec")
+            _process_transforms(files=files, data_access_factory=data_access_factory, print_interval= print_interval,
+                               statistics=statistics, runtime_configuration=runtime_config)
         status = "success"
         return_code = 0
     except Exception as e:
@@ -121,3 +103,82 @@ def orchestrate(
     except Exception as e:
         logger.error(f"Exception during execution {e}: {traceback.print_exc()}")
         return 1
+
+def _process_transforms(files: list[str], print_interval: int, data_access_factory: DataAccessFactoryBase,
+                       statistics: TransformStatistics, runtime_configuration: PythonTransformRuntimeConfiguration) \
+        -> None:
+    """
+    Process transforms sequentially
+    :param files: list of files to process
+    :param statistics: statistics class
+    :param print_interval: print interval
+    :param data_access_factory: data access factory
+    :param runtime_configuration: transform configuration
+    :return: metadata for the execution
+
+    :return: None
+    """
+    # create executor
+    executor = PythonTransformFileProcessor(
+        data_access_factory=data_access_factory, statistics=statistics, runtime_configuration=runtime_configuration
+    )
+    # process data
+    logger.debug(f"{runtime_configuration.get_name()} Begin processing files")
+    t_start = time.time()
+    completed = 0
+    for path in files:
+        executor.process_file(path)
+        completed += 1
+        if completed % print_interval == 0:
+            logger.info(
+                f"Completed {completed} files ({100 * completed / len(files)}%) "
+                f"in {(time.time() - t_start)/60} min"
+            )
+    logger.info(f"Done processing {completed} files, waiting for flush() completion.")
+    # invoke flush to ensure that all results are returned
+    start = time.time()
+    executor.flush()
+    logger.info(f"done flushing in {time.time() - start} sec")
+
+def _process_transforms_multiprocessor(files: list[str], size: int, print_interval: int,
+                                      data_access_factory: DataAccessFactoryBase,
+                                      runtime_configuration: PythonTransformRuntimeConfiguration) \
+        -> TransformStatistics:
+    """
+    Process transforms using multiprocessing pool
+    :param files: list of files to process
+    :param size: pool size
+    :param print_interval: print interval
+    :param data_access_factory: data access factory
+    :param runtime_configuration: transform configuration
+    :return: metadata for the execution
+    """
+    # result statistics
+    statistics = TransformStatistics()
+    # create processor
+    processor = PythonPoolTransformFileProcessor(data_access_factory=data_access_factory,
+                                                 runtime_configuration=runtime_configuration)
+    completed = 0
+    t_start = time.time()
+    # create multiprocessing pool
+    with Pool(processes=size) as pool:
+        # execute for every input file
+        for result in pool.imap_unordered(processor.process_file, files):
+            completed += 1
+            # accumulate statistics
+            statistics.add_stats(result)
+            if completed % print_interval == 0:
+                # print intermediate statistics
+                logger.info(
+                    f"Completed {completed} files ({100 * completed / len(files)}%) "
+                    f"in {(time.time() - t_start)/60} min"
+                )
+        logger.info(f"Done processing {completed} files, waiting for flush() completion.")
+        results = [{}] * size
+        # flush
+        for i in range(size):
+            results[i] = pool.apply_async(processor.flush)
+        for s in results:
+            statistics.add_stats(s.get())
+    logger.info(f"done flushing in {time.time() - t_start} sec")
+    return statistics
