+++ data-processing-lib/python/src/data_processing/runtime/transform_file_processor.py
@@ -14,7 +14,6 @@ import traceback
 from typing import Any
 
 from data_processing.data_access import DataAccessFactoryBase
-from data_processing.runtime import TransformRuntimeConfiguration
 from data_processing.utils import TransformUtils, UnrecoverableException, get_logger
